refactor(objectives): extract Source screening
Why:
- Confirmed Objective framing shared one service with routing, extraction, grounding, and publication, obscuring the first evidence-review decision and coupling tests to private coordinator methods.

What:
- Move Source screening records and the complete framing lifecycle to a direct analysis owner, then update the coordinator and tests to call that owner directly.

Impact:
- Framing prompts, Source ordering and identities, model budgets, progress, fallback semantics, and analysis outputs remain unchanged.

Verification:
- All 120 targeted Objective analysis tests passed, including framing, routing, extraction, comparison, and workflow behavior; package compilation and docs governance also passed.
